Move sync controller types to types module
